IndexedDB/tests/submissions/Opera/Microsoft-port/idbdatabase_createObjectStore2.htm

<!DOCTYPE html>
<title>IDBDatabase.createObjectStore() - object store 'name' and 'keyPath' properties are correctly set </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>

var t = async_test(),
    name = 'properties',
    open_rq = createdb(t)

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result,
        objStore = db.createObjectStore(name, { keyPath: "mykeypath" })

    assert_equals(objStore.name, name, "object store name")
    assert_equals(objStore.keyPath, "mykeypath", "key path")

    t.done()
}
</script>

<div id="log"></div>
